Create A28 - WAP to find 2nd largest element in an array
Description:

Read size and array elements from the user. You need to find 2nd largesst element in the array without sorting or modifying the array.
You need to pass array to the function. U need to collect as pointer inside a function as shown below:
              int sec_largest(int *arr, int size);
In function you need to find second largest element of the array and you need to return it to the main
Pre-requisites:
Arrays
Functions
Pointers

Sample Execution:
Test Case 1:
Enter the size of the Array : 5
Enter the elements into the array: 5 1 4 2 8
Second largest element of the array is 5

Test Case 2:
Enter the size of the Array : 4
Enter the elements into the array: 66 22 11 3
Second largest element of the array is 22


 Assignment slide : https://moodle.emertxe-group.com/pluginfile.php/340580/mod_vpl/intro/13_2nd_largest.pdf

Requested files
sec_largest.c

#include <stdio.h>

int sec_largest(int [], int);

int main()
{
    int size, ret;
    
    //Read size from the user
    printf("Enter the size of the array :");
    scanf("%d", &size);
    
    int arr[size];
    
    //Read elements into the array
    
    //funtion call
    ret = sec_largest(arr, size);
    
    printf("Second largest element of the array is %d\n", ret);
}